Refactor done and passess all tests
