+++ src/main/scala/es/weso/parser/TurtleParserState.scala
@@ -1,9 +1,9 @@
 package es.weso.parser
 
-import es.weso.rdfgraph.nodes._
+import es.weso.rdf.nodes._
 import scala.language.postfixOps
 import es.weso.rdf._
-import es.weso.rdfgraph.statements.RDFTriple
+import es.weso.rdf.triples.RDFTriple
 
 case class TurtleParserState(triples: List[RDFTriple], namespaces: PrefixMap, bNodeLabels: BNodeTable, baseIRI: IRI) {
