+++ src/main/scala/es/weso/rdf/RDFTriples.scala
@@ -1,13 +1,13 @@
 package es.weso.rdf
 
 /** Implementation of RDF trait using Sets of triples **/
-import es.weso.rdfgraph.statements._
+import es.weso.rdf.triples._
 // import es.weso.rdfgraph.RDFGraph
 import es.weso.parser.TurtleParser
 import scala.util._
-import es.weso.rdfgraph.nodes.InitialBNodeId
-import es.weso.rdfgraph.nodes.IRI
-import es.weso.rdfgraph.nodes.RDFNode
+import es.weso.rdf.nodes.InitialBNodeId
+import es.weso.rdf.nodes.IRI
+import es.weso.rdf.nodes.RDFNode
 
 case class RDFTriples(
     triples: Set[RDFTriple],
