+++ src/main/scala/es/weso/rdfgraph/statements/RDFTriple.scala
@@ -1,7 +1,7 @@
-package es.weso.rdfgraph.statements
+package es.weso.rdf.triples
 
 import scala.collection.Set
-import es.weso.rdfgraph.nodes._
+import es.weso.rdf.nodes._
 
 case class RDFTriple(subj: RDFNode, pred: IRI, obj: RDFNode) {
 
@@ -72,7 +72,7 @@ object RDFTriple {
    * TODO: Consider removing this method
    */
   def showTriples(triples: Set[RDFTriple]): String = {
-    var str = new StringBuilder
+    val str = new StringBuilder
     for { t <- triples } {
       str ++= (t + "\n")
     }
