+++ src/test/scala/es/weso/parser/TurtleParserSuite.scala
@@ -1,8 +1,8 @@
 package es.weso.parser
 
 import org.scalatest.{ Matchers, FunSpec }
-import es.weso.rdfgraph.nodes._
-import es.weso.rdfgraph.statements.RDFTriple
+import es.weso.rdf.nodes._
+import es.weso.rdf.triples.RDFTriple
 import es.weso.rdf.PrefixMap
 
 class TurtleParserSuite
